Extract CartLoaderCallbacks
