+++ app/src/main/java/com/github/geekarist/henripotier/CursorAdaptable.java
@@ -1,0 +1,7 @@
+package com.github.geekarist.henripotier;
+
+import android.database.Cursor;
+
+public interface CursorAdaptable {
+    void onCursorLoaded(Cursor data);
+}
